requirements.txt
flash
Flask-SQLAlchemy
pillow
python-dotenv
werkzeug
flask-login